test/LayoutTests/Panel/lastChangeSamePropertyDifferentLocations.html

<html>
<body>
<h1>Feature Test lastChange issue 1</h1>
</body>
<script src="chrome-extension://klmlfkibgfifmkanocmdenpieghpgifl/ChannelPlate/ChannelPlate.js"></script>
<script src="chrome-extension://klmlfkibgfifmkanocmdenpieghpgifl/ChannelPlate/RemoteMethodCall.js"></script>
<script src="chrome-extension://klmlfkibgfifmkanocmdenpieghpgifl/AsyncMachine.js"></script>
<script src="../../../DebugLogger.js"></script>
<script src="../PatientSelector.js"></script>
<script src="../panel-test.js"></script>

<script>

window.addEventListener('load', function() {
  testScript();
});

function test_lastChange() {
  console.log("------------test_lastChange begins------------");
  AsyncMachine.pushOp(evaluateInPage, 'window.location="http://localhost:8686/test/qpd.html"')
  AsyncMachine.pushOp(openQuerypointPanel);
  AsyncMachine.pushOp(openSourceFile, 'demo.js');

  var objPropSelector = [
    {type: 'span.cm-string', text:'\"I\'ve been clicked \"'},
    {type: 'span.cm-property', text: 'prop'}
    ];
  AsyncMachine.pushOp(selectTokenInSource, objPropSelector);
  AsyncMachine.pushOp(verifyTokenView, 'obj.prop');
  AsyncMachine.pushOp(clickQPOperation, 'lastChange');
  AsyncMachine.pushOp(whenSelectorAll, '.currentLoadNumber', '1');

  AsyncMachine.pushOp(clickSelector, '.explainTokenPanel', 'Click to Select Another');

    var anotherObjPropSelector = [
      {type: 'span.cm-variable', text:'bar'},  // call in onClick
      {type: 'span.cm-variable', text:'bar'},  // func decl
      {type: 'span.cm-property', text: 'prop'},  // obj.prop = 0
      {type: 'span.cm-property', text: 'prop'}   // prop in obj.prop = 0
    ];
  AsyncMachine.pushOp(selectTokenInSource, anotherObjPropSelector);
  AsyncMachine.pushOp(verifyTokenView, 'obj.prop');
  AsyncMachine.pushOp(clickQPOperation, 'lastChange');  

  AsyncMachine.pushOp(whenSelectorAll, '.currentLoadNumber', '2');
  AsyncMachine.pushOp(evaluateInPage, 'document.querySelector("#myButton").dispatchEvent(new MouseEvent("click"))');
  AsyncMachine.pushOp(whenSelectorAll, 'td.traceValue', '0');
  AsyncMachine.pushOp(extractText, 'td.traceValue');

  AsyncMachine.runOp();
}

var testScript = test_lastChange;
</script>
</html>
