test/LayoutTests/Panel/lastChangeSamePropertyDifferentLocations-expected.txt

http://localhost:8686/test/qpd.html
prop in   var message = "I've been clicked " + obj.prop + " times";
prop in     obj.prop = 0;
true
1|0
